Moved helper method to SignalproducerSpec
