+++ ReactiveCocoaTests/Swift/SignalProducerSpec.swift
@@ -559,3 +559,26 @@ class SignalProducerSpec: QuickSpec {
 		}
 	}
 }
+
+extension SignalProducer {
+	/// Creates a producer that can be started as many times as elements in `results`.
+	/// Each signal will immediately send either a value or an error.
+	private static func tryWithResults<C: CollectionType where C.Generator.Element == Result<T, E>, C.Index.Distance == Int>(results: C) -> SignalProducer<T, E> {
+		let resultCount = countElements(results)
+		var operationIndex = 0
+
+		precondition(resultCount > 0)
+
+		let operation: () -> Result<T, E> = {
+			if operationIndex < resultCount {
+				return results[advance(results.startIndex, operationIndex++)]
+			} else {
+				fail("Operation started too many times")
+
+				return results[advance(results.startIndex, 0)]
+			}
+		}
+
+		return SignalProducer.try(operation)
+	}
+}
